+++ src/main/java/org/example/expert/global/common/logging/AuthCheckInterceptor.java
@@ -1,4 +1,4 @@
-package org.example.expert.global.config.logging;
+package org.example.expert.global.common.logging;
 
 import io.micrometer.common.util.StringUtils;
 import jakarta.servlet.http.HttpServletRequest;
@@ -8,8 +8,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.example.expert.global.auth.JwtUtil;
 import org.example.expert.domain.auth.exception.AuthException;
 import org.example.expert.domain.user.enums.UserRole;
-import org.example.expert.global.config.logging.request.CustomHttpRequestWrapper;
-import org.example.expert.global.config.logging.response.CustomHttpResponseWrapper;
+import org.example.expert.global.common.logging.request.CustomHttpRequestWrapper;
+import org.example.expert.global.common.logging.response.CustomHttpResponseWrapper;
 import org.json.simple.JSONObject;
 import org.springframework.stereotype.Component;
 import org.springframework.web.context.request.RequestContextHolder;
