+++ src/main/java/org/example/expert/global/common/logging/request/CustomHttpRequestWrapper.java
@@ -1,4 +1,4 @@
-package org.example.expert.global.config.logging.request;
+package org.example.expert.global.common.logging.request;
 
 import jakarta.servlet.ReadListener;
 import jakarta.servlet.ServletInputStream;
